+++ kmath-ast/src/jvmMain/kotlin/space/kscience/kmath/asm/asm.kt
@@ -2,9 +2,9 @@ package space.kscience.kmath.asm
 
 import space.kscience.kmath.asm.internal.AsmBuilder
 import space.kscience.kmath.asm.internal.buildName
-import space.kscience.kmath.ast.MST
-import space.kscience.kmath.ast.MST.*
 import space.kscience.kmath.expressions.Expression
+import space.kscience.kmath.expressions.MST
+import space.kscience.kmath.expressions.MST.*
 import space.kscience.kmath.expressions.invoke
 import space.kscience.kmath.misc.Symbol
 import space.kscience.kmath.operations.Algebra
@@ -34,15 +34,17 @@ internal fun <T : Any> MST.compileWith(type: Class<T>, algebra: Algebra<T>): Exp
 
         is Unary -> when {
             algebra is NumericAlgebra && node.value is Numeric -> loadObjectConstant(
-                algebra.unaryOperationFunction(node.operation)(algebra.number(node.value.value)))
+                algebra.unaryOperationFunction(node.operation)(algebra.number((node.value as Numeric).value)))
 
             else -> buildCall(algebra.unaryOperationFunction(node.operation)) { visit(node.value) }
         }
 
         is Binary -> when {
             algebra is NumericAlgebra && node.left is Numeric && node.right is Numeric -> loadObjectConstant(
-                algebra.binaryOperationFunction(node.operation)
-                    .invoke(algebra.number(node.left.value), algebra.number(node.right.value))
+                algebra.binaryOperationFunction(node.operation).invoke(
+                    algebra.number((node.left as Numeric).value),
+                    algebra.number((node.right as Numeric).value)
+                )
             )
 
             algebra is NumericAlgebra && node.left is Numeric -> buildCall(
@@ -71,18 +73,18 @@ internal fun <T : Any> MST.compileWith(type: Class<T>, algebra: Algebra<T>): Exp
 /**
  * Create a compiled expression with given [MST] and given [algebra].
  */
-public inline fun <reified T: Any> MST.compileToExpression(algebra: Algebra<T>): Expression<T> =
+public inline fun <reified T : Any> MST.compileToExpression(algebra: Algebra<T>): Expression<T> =
     compileWith(T::class.java, algebra)
 
 
 /**
  * Compile given MST to expression and evaluate it against [arguments]
  */
-public inline fun <reified T: Any> MST.compile(algebra: Algebra<T>, arguments: Map<Symbol, T>): T =
+public inline fun <reified T : Any> MST.compile(algebra: Algebra<T>, arguments: Map<Symbol, T>): T =
     compileToExpression(algebra).invoke(arguments)
 
 /**
  * Compile given MST to expression and evaluate it against [arguments]
  */
-public inline fun <reified T: Any> MST.compile(algebra: Algebra<T>, vararg arguments: Pair<Symbol,T>): T =
+public inline fun <reified T : Any> MST.compile(algebra: Algebra<T>, vararg arguments: Pair<Symbol, T>): T =
     compileToExpression(algebra).invoke(*arguments)
